+++ previous-versions/banking/src/codatbanking/models/operations/list_transactions.py
@@ -5,7 +5,7 @@ import dataclasses
 import requests as requests_http
 from ..shared import errormessage as shared_errormessage
 from ..shared import transactions as shared_transactions
-from typing import Optional
+from typing import Callable, Optional
 
 
 
@@ -28,6 +28,7 @@ class ListTransactionsRequest:
 
 @dataclasses.dataclass
 class ListTransactionsResponse:
+    next: Callable[[], Optional[ListTransactionsResponse]] = dataclasses.field()
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)